% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{obtener_factor_rastrojo}
\alias{obtener_factor_rastrojo}
\title{Factor de reduccion de demanda evaporativa por cobertura de rastrojo}
\usage{
obtener_factor_rastrojo(constantes, clase)
}
\arguments{
\item{constantes}{Lista devuelta por \code{\link[=leer_constantes_yaml]{leer_constantes_yaml()}}.}

\item{clase}{Character. Una de \code{names(constantes$rastrojo)} (p.ej. \code{"Baja"}).}
}
\value{
Numeric (0-1).
}
\description{
Factor de reduccion de demanda evaporativa por cobertura de rastrojo
}
