% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{obtener_fraccion_humedad_inicial}
\alias{obtener_fraccion_humedad_inicial}
\title{Fraccion de agua util inicial por clase de humedad}
\usage{
obtener_fraccion_humedad_inicial(constantes, clase)
}
\arguments{
\item{constantes}{Lista devuelta por \code{\link[=leer_constantes_yaml]{leer_constantes_yaml()}}.}

\item{clase}{Character. Una de \code{names(constantes$humedad_inicial)} (p.ej. \code{"Se"}).}
}
\value{
Numeric (0-1).
}
\description{
Fraccion de agua util inicial por clase de humedad
}
