% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{obtener_latitud_estacion}
\alias{obtener_latitud_estacion}
\title{Latitud de una estacion}
\usage{
obtener_latitud_estacion(parametros, station_id)
}
\arguments{
\item{parametros}{Lista devuelta por \code{\link[=leer_parametros_yaml]{leer_parametros_yaml()}}.}

\item{station_id}{Character o numeric. Identificador de la estacion.}
}
\value{
Numeric. Latitud en grados decimales.
}
\description{
Latitud de una estacion
}
