% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{obtener_profundidad_maxima_suelo}
\alias{obtener_profundidad_maxima_suelo}
\title{Profundidad maxima de un suelo}
\usage{
obtener_profundidad_maxima_suelo(parametros, soil_id)
}
\arguments{
\item{parametros}{Lista devuelta por \code{\link[=leer_parametros_yaml]{leer_parametros_yaml()}}.}

\item{soil_id}{Character. Identificador del suelo.}
}
\value{
Numeric. Profundidad maxima del suelo, en metros.
}
\description{
Profundidad maxima de un suelo
}
